Add tags && accept followsFrom when creating the spans
Accept follows from when creating spans so we can use it instead
of childOf when the parent does not depends on the results from
the childSpan but just causes it. This is a common relationship
on producer-consumer relationship, the consumer sends a message
for the consumer to consume it, but it does not usually cares about
the result and considers its work finished once the message has
been delivered.