Add sentry release
With the release team we will start reporting releases and deployments to sentry,
telling sentry the artiifact version the code belongs to will provide us advantages
as for example detecting in which release an error was introduced, providing insight
on the releases, etc.

NOTE: Keep in mind that releases are global in sentry that's why we need to prefix the
version with the pkg.name.